+++ fast64_internal/oot/cutscene/exporter/functions.py
@@ -1,7 +1,7 @@
 import bpy
 
 from bpy.types import Object
-from .....utility import PluginError
+from ....utility import PluginError
 from .classes import OOTCSMotionExport
